HospiPlan/backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.2
django-cors-headers==4.3.1
psycopg2-binary
python-decouple==3.8
gunicorn==22.0.0
